requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
cssselect2==0.8.0
deprecation==2.1.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
fonttools==4.60.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
postgrest==2.19.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.4.2
python-dotenv==1.1.1
pytz==2025.2
realtime==2.19.0
sniffio==1.3.1
storage3==2.19.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.19.0
supabase-functions==2.19.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
Flask-Assets
jsmin
cssmin
pywebpush==2.1.0
bleach==6.1.0